requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
annotated-types
arabic-reshaper
asgiref
asn1crypto
attrs
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
cssselect2
defusedxml
distlib
Django
django-ckeditor
django-cors-headers
django-filter
django-js-asset
djangorestframework
djangorestframework-simplejwt
filelock
frozenlist

# FIXED GOOGLE AI LIBRARIES (stable versions)
google-ai-generativelanguage
google-generativeai

google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
grpcio
grpcio-status

gunicorn
html5lib
httplib2
idna
lxml
multidict
mysqlclient
oscrypto
packaging
pillow
platformdirs
propcache
proto-plus
protobuf
psycopg
psycopg-binary
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pyHanko
pyhanko-certvalidator
PyJWT
PyMySQL
pyotp
pyparsing
pypdf
python-bidi
python-decouple
python-dotenv
PyYAML
qrcode
reportlab
requests
rsa
six
sqlparse
svglib
tinycss2
tqdm
twilio
typing_extensions
tzdata
tzlocal
uritemplate
uritools
urllib3
virtualenv
virtualenvwrapper-win
webencodings
xhtml2pdf
yarl
dj-database-url
psycopg[binary]python manage.py migrate